revert SearchJSON and implement SearchEvent in Search
