+++ @types/ol-ext/control/Search.d.ts
@@ -1,6 +1,16 @@
 import ol_control_Control, {Options as ControlOptions} from 'ol/control/Control';
 import { Feature, Map as _ol_Map_ } from 'ol';
 import { Coordinate } from 'ol/coordinate';
+import BaseEvent from 'ol/events/Event';
+
+type SearchEventOnSignature<Return> = OnSignature<EventTypes, Event, Return> &
+  OnSignature<Types | 'change' | 'error' | 'propertychange', ObjectEvent, Return> &
+  OnSignature<Types | 'select', SearchEvent, Return> &
+  CombinedOnSignature<Types | EventTypes | 'select', Return>;
+
+export enum SearchEventType {
+  SELECT = 'select',
+}
 
 export interface Options extends ControlOptions{
   className?: string;
@@ -136,4 +146,13 @@ export default class Search extends ol_control_Control {
    * @return {boolean}
    */
   equalFeatures(f1: Feature, f2: Feature): boolean;
+}
+
+export class SearchEvent extends BaseEvent {
+  constructor(type: SearchEventType, search: any, reverse: boolean, coordinate: Coordinate, option?: Options);
+
+  search: any;
+  reverse: boolean;
+  coordinate: Coordinate;
+  options?: Options;
 }
